requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==1.12.1
torchvision==0.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0
opencv-python>=4.8.0
